lia-ultimate-ai/requirements.txt
## Core Framework
fastapi==0.104.0
uvicorn[standard]==0.24.0
pydantic==2.4.0

## Database & Cache
pymongo==4.6.0
redis==5.0.1
sqlalchemy==2.0.23

## AI/ML Stack
torch==2.1.0
transformers==4.35.0
sentence-transformers==2.2.2
langchain==0.0.334
openai==1.3.0

## Data Processing
pandas==2.1.3
numpy==1.24.0
scikit-learn==1.3.0
nltk==3.8.1

## APIs & Connectors
tweepy==4.14.0
google-api-python-client==2.108.0
asyncpg==0.29.0
aiohttp==3.9.0
asyncpraw==7.7.1

## Security
cryptography==41.0.7
pyjwt==2.8.0
bcrypt==4.0.1
python-dotenv==1.0.0

## Monitoring & DevOps
psutil==5.9.6
prometheus-client==0.19.0
structlog==23.2.0

## Utilities
asyncio==3.4.3
click==8.1.7
rich==13.7.0
tqdm==4.66.1
